+++ src/into_response.rs
@@ -1,7 +1,7 @@
-use hyper::{header::CONTENT_TYPE, HeaderMap};
-
-use crate::{json, status::Status, IntoPyException, Response};
+use hyper::{body::Bytes, header::CONTENT_TYPE, HeaderMap};
 use pyo3::{prelude::*, types::PyAny, Py};
+
+use crate::{cors::Cors, exceptions::*, json, status::Status, IntoPyException, Response};
 
 type Error = Box<dyn std::error::Error>;
 
@@ -58,6 +58,50 @@ impl TryFrom<(PyObject, Status)> for Response {
             headers,
             body: json::dumps(&val.0)?.into(),
         })
+    }
+}
+
+impl From<Status> for Response {
+    fn from(val: Status) -> Self {
+        let mut headers = HeaderMap::new();
+        headers.insert(CONTENT_TYPE, "application/json".parse().unwrap());
+        Response {
+            status: val,
+            headers,
+            body: Bytes::new(),
+        }
+    }
+}
+
+impl From<PyErr> for Response {
+    fn from(value: PyErr) -> Self {
+        Python::with_gil(|py| {
+            let status = match value.is_instance_of::<BaseError>(py) {
+                true if value.is_instance_of::<UnauthorizedError>(py) => Status::UNAUTHORIZED,
+                true if value.is_instance_of::<ForbiddenError>(py) => Status::FORBIDDEN,
+                true if value.is_instance_of::<NotFoundError>(py) => Status::NOT_FOUND,
+                true if value.is_instance_of::<ConflictError>(py) => Status::CONFLICT,
+                true if value.is_instance_of::<InternalError>(py) => Status::INTERNAL_SERVER_ERROR,
+                true => Status::BAD_REQUEST,
+                false => {
+                    value.display(py);
+                    Status::INTERNAL_SERVER_ERROR
+                }
+            };
+            let response: Response = status.into();
+            response.set_body(format!(
+                r#"{{"detail": "{}"}}"#,
+                value.value(py).to_string().replace('"', "'")
+            ))
+        })
+    }
+}
+
+impl From<Cors> for Response {
+    fn from(val: Cors) -> Self {
+        let mut response = Status::NO_CONTENT.into();
+        val.apply_headers(&mut response);
+        response
     }
 }
